Hot fixes an error on staging
Fixes an issue where the entry partial was also checking to see if ads should be inserted, but it wasn't using variables that existed in the partial, causing a warning. If the warning itself was fixed, too many ads showed up.

This fixes that warning, fixes a fatal that will show up in production (and on my local), and fixes the multiple ad slots issue that shows up after that.